refactor(ui): le composant statistiques globales est en tsx
